fix edit rooms

+++ java/src/com/presnakov/hotelBookingSystem/dto/room/RoomClassDto.java
@@ -4,10 +4,12 @@ import com.presnakov.hotelBookingSystem.entity.RoomClassEnum;
 import lombok.Builder;
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @Value
 @Builder
 public class RoomClassDto {
     Integer id;
     RoomClassEnum comfortClass;
-    Double pricePerDay;
+    BigDecimal pricePerDay;
 }
